Deprecate server: eliminate app.js and other files
Also renames server to api
