+++ client/src/components/RSVP.js
@@ -71,7 +71,8 @@ const RSVP = () => {
   };
 
   const sendData = async () => {
-    const url = 'https://flora-and-grahams-wedding.grahamnessler.now.sh/server/rsvpService.js';
+    const url =
+      'https://flora-and-grahams-wedding.grahamnessler.now.sh/api/rsvpService.js';
     if (!firstName || !lastName) {
       alert(invalidInputError);
       return;
