--- server/__tests__/app.test.js
@@ -1,10 +1,0 @@
-/* global it, describe, expect */
-// const App = require('../app');
-
-// TODO: create 'express' module mock and remove dummy test below
-describe('App.', () => {
-  it('should be a function', () => {
-    // expect(App).toBeInstanceOf(Function);
-    expect(true).toEqual(true);
-  });
-});
